+++ library/src/androidTest/java/cloud/pace/sdk/appkit/UriManagerTest.kt
@@ -17,7 +17,7 @@ class UriManagerTest {
 
     @Before
     fun init() {
-        PACECloudSDK.configuration = Configuration("", "", "", "", environment = Environment.DEVELOPMENT)
+        PACECloudSDK.configuration = Configuration("", "", "", "", environment = Environment.DEVELOPMENT, oidConfiguration = null)
     }
 
     @Test