+++ library/src/test/java/cloud/pace/sdk/appkit/AppRepositoryTest.kt
@@ -82,7 +82,7 @@ class AppRepositoryTest {
 
     @Before
     fun init() {
-        PACECloudSDK.configuration = Configuration("", "", "", "", environment = Environment.DEVELOPMENT)
+        PACECloudSDK.configuration = Configuration("", "", "", "", environment = Environment.DEVELOPMENT, oidConfiguration = null)
     }
 
     @Test